requirements.txt
fastapi
uvicorn
litellm
deepagents
browser-use
playwright
asyncpg
pydantic
python-dotenv
langsmith
